initial work to clean WarpX Utils
