+++ Source/Utils/Algorithms/Make.package
@@ -1,0 +1,1 @@
+VPATH_LOCATIONS   += $(WARPX_HOME)/Source/Utils/Math
