+++ Source/Utils/Make.package
@@ -11,4 +11,5 @@ CEXE_sources += MPIInitHelpers.cpp
 CEXE_sources += RelativeCellPosition.cpp
 CEXE_sources += ParticleUtils.cpp
 
+INCLUDE_LOCATIONS += $(WARPX_HOME)/Source/Utils/Math
 VPATH_LOCATIONS   += $(WARPX_HOME)/Source/Utils
